Move segments related stuff into their own module
