+++ evm_arithmetization/src/prover.rs
@@ -13,7 +13,6 @@ use plonky2::iop::challenger::Challenger;
 use plonky2::plonk::config::{GenericConfig, GenericHashOut};
 use plonky2::timed;
 use plonky2::util::timing::TimingTree;
-use serde::{Deserialize, Serialize};
 use starky::config::StarkConfig;
 use starky::cross_table_lookup::{get_ctl_data, CtlData};
 use starky::lookup::GrandProductChallengeSet;
@@ -23,39 +22,10 @@ use starky::stark::Stark;
 
 use crate::all_stark::{AllStark, Table, NUM_TABLES};
 use crate::cpu::kernel::aggregator::KERNEL;
-use crate::cpu::kernel::interpreter::{set_registers_and_run, ExtraSegmentData, Interpreter};
-use crate::generation::state::State;
-use crate::generation::{debug_inputs, generate_traces, GenerationInputs, TrimmedGenerationInputs};
+use crate::generation::segments::GenerationSegmentData;
+use crate::generation::{generate_traces, GenerationInputs, TrimmedGenerationInputs};
 use crate::get_challenges::observe_public_values;
 use crate::proof::{AllProof, MemCap, PublicValues, DEFAULT_CAP_LEN};
-use crate::witness::memory::MemoryState;
-use crate::witness::state::RegistersState;
-use crate::AllData;
-
-/// Structure holding the data needed to initialize a segment.
-#[derive(Clone, Default, Debug, Serialize, Deserialize)]
-pub struct GenerationSegmentData {
-    /// Indicates the position of this segment in a sequence of
-    /// executions for a larger payload.
-    pub(crate) segment_index: usize,
-    /// Registers at the start of the segment execution.
-    pub(crate) registers_before: RegistersState,
-    /// Registers at the end of the segment execution.
-    pub(crate) registers_after: RegistersState,
-    /// Memory at the start of the segment execution.
-    pub(crate) memory: MemoryState,
-    /// Extra data required to initialize a segment.
-    pub(crate) extra_data: ExtraSegmentData,
-    /// Log of the maximal cpu length.
-    pub(crate) max_cpu_len_log: Option<usize>,
-}
-
-impl GenerationSegmentData {
-    /// Retrieves the index of this segment.
-    pub fn segment_index(&self) -> usize {
-        self.segment_index
-    }
-}
 
 /// Generate traces, then create all STARK proofs.
 pub fn prove<F, C, const D: usize>(
@@ -482,165 +452,16 @@ pub fn check_abort_signal(abort_signal: Option<Arc<AtomicBool>>) -> Result<()> {
     Ok(())
 }
 
-/// Builds a new `GenerationSegmentData`.
-#[allow(clippy::unwrap_or_default)]
-fn build_segment_data<F: RichField>(
-    segment_index: usize,
-    registers_before: Option<RegistersState>,
-    registers_after: Option<RegistersState>,
-    memory: Option<MemoryState>,
-    interpreter: &Interpreter<F>,
-) -> GenerationSegmentData {
-    GenerationSegmentData {
-        segment_index,
-        registers_before: registers_before.unwrap_or(RegistersState::new()),
-        registers_after: registers_after.unwrap_or(RegistersState::new()),
-        memory: memory.unwrap_or(MemoryState {
-            preinitialized_segments: interpreter
-                .generation_state
-                .memory
-                .preinitialized_segments
-                .clone(),
-            ..Default::default()
-        }),
-        max_cpu_len_log: interpreter.get_max_cpu_len_log(),
-        extra_data: ExtraSegmentData {
-            bignum_modmul_result_limbs: interpreter
-                .generation_state
-                .bignum_modmul_result_limbs
-                .clone(),
-            rlp_prover_inputs: interpreter.generation_state.rlp_prover_inputs.clone(),
-            withdrawal_prover_inputs: interpreter
-                .generation_state
-                .withdrawal_prover_inputs
-                .clone(),
-            ger_prover_inputs: interpreter.generation_state.ger_prover_inputs.clone(),
-            trie_root_ptrs: interpreter.generation_state.trie_root_ptrs.clone(),
-            jumpdest_table: interpreter.generation_state.jumpdest_table.clone(),
-            next_txn_index: interpreter.generation_state.next_txn_index,
-        },
-    }
-}
-
-pub struct SegmentDataIterator<F: RichField> {
-    interpreter: Interpreter<F>,
-    partial_next_data: Option<GenerationSegmentData>,
-}
-
-pub type SegmentRunResult = Option<Box<(GenerationSegmentData, Option<GenerationSegmentData>)>>;
-
-#[derive(thiserror::Error, Debug, Serialize, Deserialize)]
-#[error("{}", .0)]
-pub struct SegmentError(pub String);
-
-impl<F: RichField> SegmentDataIterator<F> {
-    pub fn new(inputs: &GenerationInputs, max_cpu_len_log: Option<usize>) -> Self {
-        debug_inputs(inputs);
-
-        let interpreter = Interpreter::<F>::new_with_generation_inputs(
-            KERNEL.global_labels["init"],
-            vec![],
-            inputs,
-            max_cpu_len_log,
-        );
-
-        Self {
-            interpreter,
-            partial_next_data: None,
-        }
-    }
-
-    /// Returns the data for the current segment, as well as the data -- except
-    /// registers_after -- for the next segment.
-    fn generate_next_segment(
-        &mut self,
-        partial_segment_data: Option<GenerationSegmentData>,
-    ) -> Result<SegmentRunResult, SegmentError> {
-        // Get the (partial) current segment data, if it is provided. Otherwise,
-        // initialize it.
-        let mut segment_data = if let Some(partial) = partial_segment_data {
-            if partial.registers_after.program_counter == KERNEL.global_labels["halt"] {
-                return Ok(None);
-            }
-            self.interpreter
-                .get_mut_generation_state()
-                .set_segment_data(&partial);
-            self.interpreter.generation_state.memory = partial.memory.clone();
-            partial
-        } else {
-            build_segment_data(0, None, None, None, &self.interpreter)
-        };
-
-        let segment_index = segment_data.segment_index;
-
-        // Run the interpreter to get `registers_after` and the partial data for the
-        // next segment.
-        let run = set_registers_and_run(segment_data.registers_after, &mut self.interpreter);
-        if let Ok((updated_registers, mem_after)) = run {
-            let partial_segment_data = Some(build_segment_data(
-                segment_index + 1,
-                Some(updated_registers),
-                Some(updated_registers),
-                mem_after,
-                &self.interpreter,
-            ));
-
-            segment_data.registers_after = updated_registers;
-            Ok(Some(Box::new((segment_data, partial_segment_data))))
-        } else {
-            let inputs = &self.interpreter.get_generation_state().inputs;
-            let block = inputs.block_metadata.block_number;
-            let txn_range = match inputs.txn_hashes.len() {
-                0 => "Dummy".to_string(),
-                1 => format!("{:?}", inputs.txn_number_before),
-                _ => format!(
-                    "{:?}_{:?}",
-                    inputs.txn_number_before,
-                    inputs.txn_number_before + inputs.txn_hashes.len()
-                ),
-            };
-            let s = format!(
-                "Segment generation {:?} for block {:?} ({}) failed with error {:?}",
-                segment_index,
-                block,
-                txn_range,
-                run.unwrap_err()
-            );
-            Err(SegmentError(s))
-        }
-    }
-}
-
-impl<F: RichField> Iterator for SegmentDataIterator<F> {
-    type Item = AllData;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let run = self.generate_next_segment(self.partial_next_data.clone());
-
-        if let Ok(segment_run) = run {
-            match segment_run {
-                // The run was valid, but didn't not consume the payload fully.
-                Some(boxed) => {
-                    let (data, next_data) = *boxed;
-                    self.partial_next_data = next_data;
-                    Some(Ok((self.interpreter.generation_state.inputs.clone(), data)))
-                }
-                // The payload was fully consumed.
-                None => None,
-            }
-        } else {
-            // The run encountered some error.
-            Some(Err(run.unwrap_err()))
-        }
-    }
-}
-
 /// A utility module designed to test witness generation externally.
 pub mod testing {
     use super::*;
     use crate::{
         cpu::kernel::interpreter::Interpreter,
-        generation::{output_debug_tries, state::State},
+        generation::{
+            output_debug_tries,
+            segments::{SegmentDataIterator, SegmentError},
+            state::State,
+        },
     };
 
     /// Simulates the zkEVM CPU execution.
@@ -677,7 +498,8 @@ pub mod testing {
         let mut proofs = vec![];
 
         for segment_run in segment_data_iterator {
-            let (_, mut next_data) = segment_run.map_err(|e| anyhow::format_err!(e))?;
+            let (_, mut next_data) =
+                segment_run.map_err(|e: SegmentError| anyhow::format_err!(e))?;
             let proof = prove(
                 all_stark,
                 config,
